+++ data/src/main/java/ru/kima/intelligentchat/data/kobold/horde/model/RequestAsyncDto.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class RequestAsyncDto(
     @SerialName("id") val id: String = String(),
-    @SerialName("kudos") val kudos: Int = 0,
+    @SerialName("kudos") val kudos: Double = 0.0,
     @SerialName("message") val message: String = String(),
     @SerialName("warnings") val warnings: List<RequestSingleWarningDto> = emptyList()
 )
