+++ domain/src/main/java/ru/kima/intelligentchat/domain/horde/model/HordeAsyncRequest.kt
@@ -2,7 +2,7 @@ package ru.kima.intelligentchat.domain.horde.model
 
 data class HordeAsyncRequest(
     val id: String = String(),
-    val kudos: Int = 0,
+    val kudos: Double = 0.0,
     val message: String = String(),
     val warnings: List<String> = emptyList()
 )
